requirements_evaluation.txt
requests>=2.31.0
aiohttp>=3.8.0
google-generativeai>=0.3.0
pandas>=2.0.0
sqlite3
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.28.0
numpy>=1.24.0
asyncio 